feat(loadgen): render the failure reason into sampled NOK callflows
A sampled failing call previously showed only the FAIL status with no
explanation. Thread the call's outcome detail (the StepError / reason string)
through `render_html` → `seq_doc` → `sip_doc`: on a not-passed call it becomes
the page description (the banner next to FAIL) AND an explicit `call-result`
anomaly in the list. So a sampled NOK callflow now says WHY — e.g.
"alice expected 200, got 486 Busy Here", "transfer declined by charlie (603)" —
not just that it failed. Also rename `AsEmergency::new`→`wrap` (clippy
new_ret_no_self) and drop an unnecessary cast in the overload test.

+++ crates/loadgen/tests/smoke.rs
@@ -365,7 +365,7 @@ async fn loadgen_mux_emergency_split_under_overload() {
     // lifecycle `assert_fully_reaped`, exactly as the authoritative
     // `tier3_admission_gate` test does).
     assert!(
-        b2bua.metrics().overload_rejected_total() as u64 >= ne_503,
+        b2bua.metrics().overload_rejected_total() >= ne_503,
         "every non-emergency shed must be a stateless overload reject:\n{}",
         reporter.render_prometheus()
     );
